Updated workspace for OSX port. Fixed gcc warning.

+++ src/window_menustatus.cpp
@@ -51,7 +51,7 @@ void Window_MenuStatus::Refresh() {
 	item_max = Game_Party::GetActors().size();
 
 	int y = 0;
-	for (size_t i = 0; i < item_max; ++i)
+	for (int i = 0; i < item_max; ++i)
 	{
 		DrawActorGraphic(Game_Party::GetActors()[i], 0, i*48 + y);
 		DrawActorName(Game_Party::GetActors()[i], 48 + 8, i*48 + 2 + y);
